requirements.txt
streamlit
uvicorn
fastapi
chromadb
qdrant-client
sentence-transformers
rank-bm25
python-docx
spacy
pypdf
nltk
google-cloud-aiplatform
google-cloud-storage
google-cloud-firestore
google-cloud-pubsub
google-cloud-secret-manager
pandas
scikit-learn
transformers
FlagEmbedding
huggingface_hub
ragas
hf_xet
gunicorn
python-multipart
requests
cryptography
pytest
pytest-cov